Backend/app/requirements.txt
flask
flask-cors
google-generativeai
google-genai
python-dotenv
grpcio
gunicorn
gevent
flask_socketio